new component UserCard, fixed styles
