+++ front/src/components/PlayerSearch/PlayerSearch.tsx
@@ -1,37 +1,13 @@
 "use client";
 
 import React, { useState, useEffect, useContext, useRef } from "react";
-import { UserContext } from "../Context/UserContext";
 import TranslationContext from "../Context/TranslationContext";
 import axios from "axios";
 import { toast } from "react-toastify";
 import { useRouter } from "next/navigation";
-import Link from "next/link";
-import Image from "next/image";
-import { getDefaultPlayerImage } from "@/helpers/imageUtils";
-import { renderCountryFlag } from "../countryFlag/countryFlag";
-import { FaShieldAlt } from "react-icons/fa";
-
-interface User {
-  id: string;
-  name: string;
-  lastname: string;
-  imgUrl?: string;
-  age?: number;
-  nationality?: string;
-  primaryPosition?: string;
-  secondaryPosition?: string;
-  height?: number;
-  weight?: number;
-  skillfulFoot?: string;
-  subscriptionType?: string;
-  subscription?: string;
-  role?: string;
-  email?: string;
-  phone?: string;
-  ubicacionActual?: string;
-  trayectorias?: [];
-}
+import UserCard from "./UserCard";
+import { User } from "@/Interfaces/IUser";
+import { useUserContext } from "@/hook/useUserContext";
 
 interface SearchFilters {
   name?: string;
@@ -48,7 +24,7 @@ interface SearchFilters {
 }
 
 const PlayerSearch: React.FC = () => {
-  const { user, token } = useContext(UserContext);
+  const { user, token } = useUserContext();
   const {
     setLanguage,
     currentLanguage,
@@ -818,392 +794,14 @@ const PlayerSearch: React.FC = () => {
 
   // Renderizar tarjeta de usuario
   const renderUserCard = (currentUser: User) => {
-    const isPlayer = currentUser.role === "PLAYER";
-    const isRecruiter = currentUser.role === "RECRUITER";
-
-    // Función auxiliar para verificar si el usuario tiene una suscripción específica
-    const hasSubscription = (subscriptionName: string): boolean => {
-      return (
-        currentUser.subscriptionType === subscriptionName ||
-        currentUser.subscription === subscriptionName
-      );
-    };
-
-    // Determinar tipo de suscripción usando ambos campos: subscriptionType y subscription
-    const subscriptionType = (() => {
-      if (
-        hasSubscription("Profesional") ||
-        hasSubscription("profesional") ||
-        hasSubscription("Professional")
-      ) {
-        return "professional";
-      }
-      if (
-        hasSubscription("Semiprofesional") ||
-        hasSubscription("semiprofesional") ||
-        hasSubscription("Semi-profesional")
-      ) {
-        return "semi";
-      }
-      return "amateur";
-    })();
-
-    // Color de la insignia según tipo de suscripción o rol
-    const badgeBgClass = isRecruiter
-      ? "bg-purple-800"
-      : {
-          professional: "bg-green-800",
-          semi: "bg-blue-700",
-          amateur: "bg-gray-700",
-        }[subscriptionType];
-
-    // Texto de la insignia según tipo de suscripción o rol
-    const badgeText = isRecruiter
-      ? t("recruiter")
-      : {
-          professional: t("professional"),
-          semi: t("semi"),
-          amateur: t("amateur"),
-        }[subscriptionType];
-
-    // Verificar si este usuario está siendo añadido a la cartera
-    const isBeingAddedToPortfolio = isAddingToPortfolio === currentUser.id;
-
-    // Obtener el nombre del ultimo Club del Usuario
-    const trayectorias = currentUser?.trayectorias as
-      | { club: string }[]
-      | undefined;
-
-    const lastClub = trayectorias?.length
-      ? trayectorias[trayectorias.length - 1].club
-      : "----";
-
     return (
-      <div
+      <UserCard
+        currentUser={currentUser}
+        t={t}
+        isAddingToPortfolio={isAddingToPortfolio}
+        handleAddToPortfolio={handleAddToPortfolio}
         key={currentUser.id}
-        className="bg-white rounded-lg overflow-hidden shadow-md hover:shadow-xl transition-all relative"
-      >
-        {/* Imagen de fondo/header */}
-        <div className="h-24 bg-gradient-to-r from-gray-200 to-gray-300 relative overflow-hidden">
-          {/* Imagen de fondo (misma que la de perfil) */}
-          {currentUser.imgUrl && (
-            <div className="absolute inset-0">
-              <Image
-                src={currentUser.imgUrl}
-                alt=""
-                fill
-                className="object-cover blur-[1px]"
-              />
-              {/* Overlay para oscurecer la imagen */}
-              <div className="absolute inset-0 bg-black/30"></div>
-            </div>
-          )}
-
-          {/* Gradiente decorativo si no hay imagen */}
-          {!currentUser.imgUrl && (
-            <div
-              className={`absolute inset-0 bg-gradient-to-r ${
-                isRecruiter
-                  ? "from-purple-800 to-purple-600"
-                  : subscriptionType === "professional"
-                  ? "from-green-800 to-green-600"
-                  : subscriptionType === "semi"
-                  ? "from-blue-700 to-blue-500"
-                  : "from-gray-700 to-gray-500"
-              }`}
-            ></div>
-          )}
-        </div>
-
-        {/* Contenido principal */}
-        <div className="px-4 pb-4 pt-12 relative">
-          {/* Foto de perfil */}
-          <div className="absolute -top-10 left-4">
-            <div className="relative">
-              <div className="w-20 h-20 rounded-full border-4 border-white overflow-hidden shadow-md">
-                <Image
-                  src={currentUser.imgUrl || getDefaultPlayerImage()}
-                  alt={`${currentUser.name} ${currentUser.lastname}`}
-                  width={80}
-                  height={80}
-                  className="object-cover w-full h-full"
-                />
-              </div>
-
-              {/* Indicador de tipo de suscripción */}
-              <div
-                className={`absolute -right-1 -bottom-1 ${badgeBgClass} text-white text-xs py-0.5 px-2 rounded-full font-medium shadow-sm`}
-              >
-                <span>{badgeText}</span>
-              </div>
-            </div>
-          </div>
-
-          {/* Nombre y información */}
-          <div className="mb-2">
-            <h3 className="text-lg font-bold text-gray-900 leading-tight">
-              {currentUser.name} {currentUser.lastname}
-            </h3>
-            <p className="text-sm text-gray-600 line-clamp-2">
-              {isPlayer && (
-                <>
-                  {currentUser.primaryPosition || t("notSpecified")}
-                  {currentUser.secondaryPosition &&
-                    ` / ${currentUser.secondaryPosition}`}
-                </>
-              )}
-              {isRecruiter && (
-                <>
-                  {t("recruiter")}
-                  {currentUser.ubicacionActual &&
-                    ` - ${currentUser.ubicacionActual}`}
-                </>
-              )}
-              {currentUser.nationality && ` | ${currentUser.nationality}`}
-            </p>
-          </div>
-
-          {/* Detalles adicionales */}
-          <div className="text-xs text-gray-500 space-y-1 mb-4">
-            {isPlayer && (
-              <>
-                {currentUser.age && (
-                  <div className="flex items-center gap-1">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="h-3.5 w-3.5"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z"
-                      />
-                    </svg>
-                    <span>
-                      {t("age")}: {currentUser.age}
-                    </span>
-                  </div>
-                )}
-
-                {lastClub && (
-                  <div className="flex items-center text-xs gap-1 text-gray-500 space-y-1 mb-4">
-                    <FaShieldAlt className="w-3.5 h-3.5" />
-                    <span>
-                      {t("Club actual")}: {lastClub}
-                    </span>
-                  </div>
-                )}
-
-                {currentUser.height && (
-                  <div className="flex items-center gap-1">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="h-3.5 w-3.5"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M5 3v4M3 5h4M6 17v4m-2-2h4m5-16l2.286 6.857L21 12l-5.714 2.143L13 21l-2.286-6.857L5 12l5.714-2.143L13 3z"
-                      />
-                    </svg>
-                    <span>
-                      {t("height")}: {currentUser.height} cm
-                    </span>
-                  </div>
-                )}
-
-                {currentUser.skillfulFoot && (
-                  <div className="flex items-center gap-1">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="h-3.5 w-3.5"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M14.828 14.828a4 4 0 01-5.656 0M9 10h.01M15 10h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-                      />
-                    </svg>
-                    <span>
-                      {t("skillfulFoot")}: {currentUser.skillfulFoot}
-                    </span>
-                  </div>
-                )}
-              </>
-            )}
-
-            {isRecruiter && (
-              <>
-                <div className="bg-yellow-50 border border-yellow-100 rounded p-1 mb-1">
-                  <div className="flex items-center">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="h-3 w-3 text-yellow-600 mr-1"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M12 15v2m0-6v4m0-8v.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-                      />
-                    </svg>
-                    <span className="text-xs text-yellow-700">
-                      Contacto no disponible con tu suscripción
-                    </span>
-                  </div>
-                </div>
-
-                <div className="flex items-center gap-1">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    className="h-3.5 w-3.5 text-gray-400"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M3 8l7.89 4.26a2 2 0 002.22 0L21 8M5 19h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"
-                    />
-                  </svg>
-                  <span className="truncate text-gray-400">●●●●●●●●</span>
-                </div>
-
-                <div className="flex items-center gap-1">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    className="h-3.5 w-3.5 text-gray-400"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M3 5a2 2 0 012-2h3.28a1 1 0 01.948.684l1.498 4.493a1 1 0 01-.502 1.21l-2.257 1.13a11.042 11.042 0 005.516 5.516l1.13-2.257a1 1 0 011.21-.502l4.493 1.498a1 1 0 01.684.949V19a2 2 0 01-2 2h-1C9.716 21 3 14.284 3 6V5z"
-                    />
-                  </svg>
-                  <span className="text-gray-400">●●●●●●●●</span>
-                </div>
-              </>
-            )}
-
-            {/* Nacionalidad con bandera */}
-            {currentUser.nationality && (
-              <div className="flex items-center gap-1">
-                <span className="flex items-center">
-                  {renderCountryFlag(currentUser.nationality)}
-                </span>
-                <span>{currentUser.nationality}</span>
-              </div>
-            )}
-          </div>
-
-          {/* Botones de acción */}
-          <div className="flex flex-col space-y-2">
-            {/* Botón de ver perfil */}
-            <Link
-              href={`/user-viewer/${currentUser.id}`}
-              className={`flex items-center justify-center w-full py-1.5 px-3 border rounded-full text-sm font-medium transition-colors ${
-                isRecruiter
-                  ? "border-purple-600 text-purple-600 hover:bg-purple-50"
-                  : "border-blue-600 text-blue-600 hover:bg-blue-50"
-              }`}
-            >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-4 w-4 mr-1"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M18 9v3m0 0v3m0-3h3m-3 0h-3m-2-5a4 4 0 11-8 0 4 4 0 018 0zM3 20a6 6 0 0112 0v1H3v-1z"
-                />
-              </svg>
-              {t("viewProfile")}
-            </Link>
-
-            {/* Botón de añadir a cartera (solo visible para reclutadores y para jugadores) */}
-            {user && user.role === "RECRUITER" && isPlayer && (
-              <button
-                onClick={(e) => {
-                  e.preventDefault();
-                  e.stopPropagation();
-                  handleAddToPortfolio(currentUser.id);
-                }}
-                disabled={isBeingAddedToPortfolio}
-                className="flex items-center justify-center w-full py-1.5 px-3 border border-green-600 text-green-600 rounded-full text-sm font-medium hover:bg-green-50 transition-colors"
-              >
-                {isBeingAddedToPortfolio ? (
-                  <span className="flex items-center">
-                    <svg
-                      className="animate-spin -ml-1 mr-2 h-4 w-4 text-green-600"
-                      xmlns="http://www.w3.org/2000/svg"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                    >
-                      <circle
-                        className="opacity-25"
-                        cx="12"
-                        cy="12"
-                        r="10"
-                        stroke="currentColor"
-                        strokeWidth="4"
-                      ></circle>
-                      <path
-                        className="opacity-75"
-                        fill="currentColor"
-                        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                      ></path>
-                    </svg>
-                    Enviando...
-                  </span>
-                ) : (
-                  <>
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="h-4 w-4 mr-1"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M8 10h.01M12 10h.01M16 10h.01M9 16H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-5l-5 5v-5z"
-                      />
-                    </svg>
-                    Enviar solicitud
-                  </>
-                )}
-              </button>
-            )}
-          </div>
-        </div>
-      </div>
+      />
     );
   };
