+++ front/src/hook/useUserContext.ts
@@ -1,0 +1,4 @@
+import { UserContext } from "@/components/Context/UserContext";
+import { useContext } from "react";
+
+export const useUserContext = () => useContext(UserContext);
